+++ Changelog.md
@@ -1,11 +1,14 @@
+### version 1.1.4
+- [X] ToHtmlTableByProperties Add ExpressionCaches , it improve the efficiency after the cache created.
+- [X] Support Cunstom Table/TR/TD Attributes.
+
 ### version 1.1.3
-- [X] Configurable InnerHtml Encoding
+- [X] Configurable InnerHtml Encoding 
 
 ### version 1.1.2
 - [X] Defalut html encode prevent xss
 
 ### version 1.1.1
-
 - [X] Support Dapper Dynamic Query
 - [X] Support Dictionary
 - [X] Support DataTable